pyproject.toml
[tool.poetry]
name = "prophet"
version = "0.1.0"
description = ""
authors = ["Raghu Reddy <raghureddy.ram@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sentence-transformers = "^2.2.2"
argparse = "^1.4.0"
text-generation = "^0.6.1"
pdfminer = "^20191125"
faiss-cpu = "^1.7.4"
pdfminer-six = "^20221105"
openai = "^1.5.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.2"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"

[tool.poetry.scripts]
earnings-reviewer = "prophet.runner:start"
earnings-server = "prophet.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
